+++ src/Akeneo/Component/FileStorage/Model/File.php
@@ -192,4 +192,12 @@ class File implements FileInterface
 
         return $this;
     }
+
+    /**
+     * @return string
+     */
+    public function __toString()
+    {
+        return $this->getOriginalFilename();
+    }
 }