+++ src/Pim/Bundle/BaseConnectorBundle/Archiver/FileReaderArchiver.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\BaseConnectorBundle\Archiver;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Item\ItemReaderInterface;
 use Akeneo\Bundle\BatchBundle\Step\ItemStep;
-use Gaufrette\Filesystem;
+use League\Flysystem\Filesystem;
 use Pim\Bundle\BaseConnectorBundle\Reader\File\FileReader;
 use Pim\Component\Connector\Reader\File\CsvReader;
 
@@ -46,7 +46,7 @@ class FileReaderArchiver extends AbstractFilesystemArchiver
                         '%filename%' => basename($reader->getFilePath()),
                     ]
                 );
-                $this->filesystem->write($key, file_get_contents($reader->getFilePath()), true);
+                $this->filesystem->put($key, file_get_contents($reader->getFilePath()));
             }
         }
     }