+++ src/Pim/Bundle/BaseConnectorBundle/Archiver/InvalidItemsCsvArchiver.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\BaseConnectorBundle\Archiver;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Gaufrette\Filesystem;
+use League\Flysystem\Filesystem;
 use Pim\Bundle\BaseConnectorBundle\EventListener\InvalidItemsCollector;
 use Pim\Bundle\BaseConnectorBundle\Writer\File\CsvWriter;
 
@@ -22,27 +22,21 @@ class InvalidItemsCsvArchiver extends AbstractFilesystemArchiver
     /** @var CsvWriter */
     protected $writer;
 
-    /** @var string */
-    protected $archiveDir;
-
     /**
      * Constructor
      *
      * @param InvalidItemsCollector $collector
      * @param CsvWriter             $writer
      * @param Filesystem            $filesystem
-     * @param string                $archiveDir
      */
     public function __construct(
         InvalidItemsCollector $collector,
         CsvWriter $writer,
-        Filesystem $filesystem,
-        $archiveDir
+        Filesystem $filesystem
     ) {
         $this->collector  = $collector;
         $this->writer     = $writer;
         $this->filesystem = $filesystem;
-        $this->archiveDir = $archiveDir;
     }
 
     /**
@@ -57,13 +51,11 @@ class InvalidItemsCsvArchiver extends AbstractFilesystemArchiver
             $this->getRelativeArchivePath($jobExecution),
             array('%filename%' => 'invalid_items.csv')
         );
-        $this->filesystem->write($key, '', true);
+        $this->filesystem->put($key, '');
         $this->writer->setFilePath(
-            sprintf(
-                '%s/%s',
-                $this->archiveDir,
-                $key
-            )
+            $this->filesystem->getConfig()->get('directory') .
+            DIRECTORY_SEPARATOR .
+            $key
         );
         $this->writer->initialize();
         $this->writer->write($this->collector->getInvalidItems());
@@ -94,5 +86,7 @@ class InvalidItemsCsvArchiver extends AbstractFilesystemArchiver
 
             return true;
         }
+
+        return false;
     }
 }